+++ snapshot-tests-directory-params/src/main/java/de/skuzzle/test/snapshots/directoryparams/FilesFromArgumentsProvider.java
@@ -3,7 +3,6 @@ package de.skuzzle.test.snapshots.directoryparams;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.Arrays;
 import java.util.Comparator;
 import java.util.stream.Stream;
 
@@ -11,19 +10,23 @@ import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.ArgumentsProvider;
 import org.junit.jupiter.params.support.AnnotationConsumer;
+import org.junit.platform.commons.support.ReflectionSupport;
 
 import de.skuzzle.test.snapshots.io.DirectoryResolver;
 
-class DirectoryContentsArgumentsProvider implements ArgumentsProvider, AnnotationConsumer<FilesFrom> {
+class FilesFromArgumentsProvider implements ArgumentsProvider, AnnotationConsumer<FilesFrom> {
 
     private FilesFrom directoryContents;
 
     @Override
     public Stream<? extends Arguments> provideArguments(ExtensionContext context) throws Exception {
         final Path inputFileDirectory = determineDirectory().toAbsolutePath();
+        final PathFilter filter = PathFilter.fromPredicate(Files::isRegularFile)
+                .and(PathFilterExtensions.extensions(directoryContents.extensions()))
+                .and(additionalFilter());
 
         return streamFiles(inputFileDirectory)
-                .filter(this::includeFile)
+                .filter(filter.toPredicate())
                 .map(TestFile::new)
                 .sorted(Comparator.comparing(TestFile::name))
                 .map(Arguments::of);
@@ -35,18 +38,8 @@ class DirectoryContentsArgumentsProvider implements ArgumentsProvider, Annotatio
                 : Files.list(root);
     }
 
-    private boolean includeFile(Path file) {
-        return directoryContents.extensions().length == 0 ||
-                Arrays.stream(directoryContents.extensions())
-                        .anyMatch(configuredExtension -> matchesExtension(configuredExtension, file));
-    }
-
-    private boolean matchesExtension(String configuredExtension, Path file) {
-        final String normalizedExtension = configuredExtension.startsWith(".")
-                ? configuredExtension
-                : "." + configuredExtension;
-        final String filesExtension = FileExtension.includingLeadingDot(file);
-        return filesExtension.equalsIgnoreCase(normalizedExtension);
+    private PathFilter additionalFilter() {
+        return ReflectionSupport.newInstance(directoryContents.filter());
     }
 
     private Path determineDirectory() throws IOException {